scraper/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic>=2.10
python-multipart==0.0.12
pypdf==3.17.4
requests==2.31.0
beautifulsoup4==4.12.3
python-docx==0.8.11
